Create model-predict
Make predictions with your created model. In this case, we will predict the avg_temperature for 150 data points (150 days).

If we changed the function "ML.FORECAST" by "ML.FORECAST_EXPLAIN", the query will return a superset of this forecast, where we will better understand our forecast. E.g. we can see that the -nonsense- detected weekly seasonality is negligible (as opposed to the yearly one).